requirements.txt
flask==2.0.1
slack-sdk==3.19.5
python-dotenv==0.19.0
requests==2.28.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
gunicorn==20.1.0 